Create new files to modularize code
